refactor: extract methods related to parsing http query params
methods were moved from GenericDataController to new abstract class
they were made protected, too
